Move FlushTasks to testsuite package and use elsewhere
