+++ core/tasks/ivr/start_ivr_flow_batch_test.go
@@ -2,7 +2,6 @@ package ivr_test
 
 import (
 	"context"
-	"encoding/json"
 	"net/http"
 	"testing"
 
@@ -11,13 +10,10 @@ import (
 	"github.com/nyaruka/gocommon/urns"
 	"github.com/nyaruka/mailroom/core/ivr"
 	"github.com/nyaruka/mailroom/core/models"
-	"github.com/nyaruka/mailroom/core/queue"
-	ivrtasks "github.com/nyaruka/mailroom/core/tasks/ivr"
 	"github.com/nyaruka/mailroom/core/tasks/starts"
 	"github.com/nyaruka/mailroom/runtime"
 	"github.com/nyaruka/mailroom/testsuite"
 	"github.com/nyaruka/mailroom/testsuite/testdata"
-
 	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 )
@@ -43,29 +39,33 @@ func TestIVR(t *testing.T) {
 	err := starts.CreateFlowBatches(ctx, rt, start)
 	assert.NoError(t, err)
 
-	// should have one task in our ivr queue
-	task, err := queue.PopNextTask(rc, queue.HandlerQueue)
-	assert.NoError(t, err)
-	batch := &models.FlowStartBatch{}
-	err = json.Unmarshal(task.Task, batch)
-	assert.NoError(t, err)
-
 	service.callError = errors.Errorf("unable to create call")
-	err = ivrtasks.HandleFlowStartBatch(ctx, rt, batch)
-	assert.NoError(t, err)
+
+	// should have one IVR batch start task
+	taskCounts := testsuite.FlushTasks(t, rt)
+	assert.Equal(t, 1, taskCounts["start_ivr_flow_batch"])
 	assertdb.Query(t, db, `SELECT COUNT(*) FROM ivr_call WHERE contact_id = $1 AND status = $2`, testdata.Cathy.ID, models.CallStatusFailed).Returns(1)
 
+	// re-create the batch and try again
 	service.callError = nil
 	service.callID = ivr.CallID("call1")
-	err = ivrtasks.HandleFlowStartBatch(ctx, rt, batch)
+
+	err = starts.CreateFlowBatches(ctx, rt, start)
 	assert.NoError(t, err)
+
+	taskCounts = testsuite.FlushTasks(t, rt)
+	assert.Equal(t, 1, taskCounts["start_ivr_flow_batch"])
 	assertdb.Query(t, db, `SELECT COUNT(*) FROM ivr_call WHERE contact_id = $1 AND status = $2 AND external_id = $3`, testdata.Cathy.ID, models.CallStatusWired, "call1").Returns(1)
 
 	// trying again should put us in a throttled state (queued)
 	service.callError = nil
 	service.callID = ivr.CallID("call1")
-	err = ivrtasks.HandleFlowStartBatch(ctx, rt, batch)
+
+	err = starts.CreateFlowBatches(ctx, rt, start)
 	assert.NoError(t, err)
+
+	taskCounts = testsuite.FlushTasks(t, rt)
+	assert.Equal(t, 1, taskCounts["start_ivr_flow_batch"])
 	assertdb.Query(t, db, `SELECT COUNT(*) FROM ivr_call WHERE contact_id = $1 AND status = $2 AND next_attempt IS NOT NULL;`, testdata.Cathy.ID, models.CallStatusQueued).Returns(1)
 }
