+++ testsuite/testsuite.go
@@ -7,14 +7,10 @@ import (
 	"os/exec"
 	"path"
 	"strings"
-	"testing"
 
-	"github.com/nyaruka/gocommon/jsonx"
 	"github.com/nyaruka/gocommon/storage"
 	"github.com/nyaruka/mailroom/core/models"
-	"github.com/nyaruka/mailroom/core/queue"
 	"github.com/nyaruka/mailroom/runtime"
-	"github.com/stretchr/testify/require"
 
 	"github.com/gomodule/redigo/redis"
 	"github.com/jmoiron/sqlx"
@@ -283,31 +279,4 @@ func ReadFile(path string) []byte {
 	d, err := os.ReadFile(path)
 	noError(err)
 	return d
-}
-
-func CurrentOrgTasks(t *testing.T, rp *redis.Pool) map[models.OrgID][]*queue.Task {
-	rc := rp.Get()
-	defer rc.Close()
-
-	// get all active org queues
-	active, err := redis.Ints(rc.Do("ZRANGE", "batch:active", 0, -1))
-	require.NoError(t, err)
-
-	tasks := make(map[models.OrgID][]*queue.Task)
-	for _, orgID := range active {
-		orgTasksEncoded, err := redis.Strings(rc.Do("ZRANGE", fmt.Sprintf("batch:%d", orgID), 0, -1))
-		require.NoError(t, err)
-
-		orgTasks := make([]*queue.Task, len(orgTasksEncoded))
-
-		for i := range orgTasksEncoded {
-			task := &queue.Task{}
-			jsonx.MustUnmarshal([]byte(orgTasksEncoded[i]), task)
-			orgTasks[i] = task
-		}
-
-		tasks[models.OrgID(orgID)] = orgTasks
-	}
-
-	return tasks
 }
